+++ project/BuildHelper.scala
@@ -10,12 +10,12 @@ object BuildHelper {
   lazy val dependencies = new {
     val betterFiles     = "com.github.pathikrit" %% "better-files"      % "3.9.2"
     val commonsCodec    = "commons-codec"        %  "commons-codec"     % "1.17.0"
+    val chimney         = "io.scalaland"         %% "chimney"           % "1.3.0"
     val firebase        = "com.google.firebase"  %  "firebase-admin"    % "9.3.0"
     val guara           = "guara"                %% "guara-framework"   % "v0.1.0-SNAPSHOT" changing()
     val jjwtApi         = "io.jsonwebtoken"      % "jjwt-api"           % "0.12.3"
     val jjwtImpl        = "io.jsonwebtoken"      % "jjwt-impl"          % "0.12.3"
     val jjwtJackson     = "io.jsonwebtoken"      % "jjwt-jackson"       % "0.12.3"
-    val chimney         = "io.scalaland"         %% "chimney"           % "0.8.2"
     val quillZio        = "io.getquill"          %% "quill-zio"         % "4.8.5"
     val quillZioJdbc    = "io.getquill"          %% "quill-jdbc-zio"    % "4.8.4"
     val postgresql      = "org.postgresql"       %  "postgresql"        % "42.7.3"
@@ -31,7 +31,8 @@ object BuildHelper {
     dependencies.zio,
     dependencies.zioJson,
     dependencies.zioOptics,
-    dependencies.guara
+    dependencies.guara,
+    dependencies.chimney
   )
 
   lazy val clientDependencies = Seq(
@@ -67,7 +68,7 @@ object BuildHelper {
     ThisBuild / scalaVersion                 := ScalaVersion,
     ThisBuild / scalacOptions                := commonSettings(scalaVersion.value),
     ThisBuild / organization                 := "morbid",
-    ThisBuild / version                      := "v0.0.2",
+    ThisBuild / version                      := "v0.0.3-SNAPSHOT",
     ThisBuild / doc / sources                := Seq.empty,
     ThisBuild / packageDoc / publishArtifact := false,
     ThisBuild / resolvers                    += Resolver.mavenLocal,